TXT Copy over section from thesis
This is a good start, but it is too succint. Also, in the thesis this
was just a section in a larger chapter. Here it needs to be broken up.
Also, it needs to be integrated into the intellectual framework of the
paper instead of just as an appendix to something unrelated.